Reorganize lasses.
